requirements.txt
opencv-python
numpy
pytesseract
spacy
scikit-learn
streamlit
pillow